TT#44197 Support Software RAID for installation
Assuming we have two disks /dev/sda + /dev/sdb and boot the
system with the (new) boot options "swraiddisk1=sda
swraiddisk2=sdb", then we will get a SW-RAID (Software RAID)
setup which looks like (the /boot/efi is only present with EFI
support and only present for the active partition):

| NAME                MAJ:MIN RM  SIZE RO TYPE  MOUNTPOINT
| sda                   8:0    0   16G  0 disk
| |-sda1                8:1    0    1M  0 part
| |-sda2                8:2    0  486M  0 part  /boot/efi
| `-sda3                8:3    0 15.5G  0 part
|   `-md0               9:0    0 15.5G  0 raid1
|     |-ngcp-root     253:0    0    5G  0 lvm   /
|     |-ngcp-fallback 253:1    0    5G  0 lvm
|     `-ngcp-data     253:3    0    5G  0 lvm   /ngcpdata
| sdb                   8:16   0   16G  0 disk
| |-sdb1                8:17   0    1M  0 part
| |-sdb2                8:18   0  486M  0 part
| `-sdb3                8:19   0 15.5G  0 part
|   `-md0               9:0    0 15.5G  0 raid1
|     |-ngcp-root     253:0    0    5G  0 lvm   /
|     |-ngcp-fallback 253:1    0    5G  0 lvm
|     `-ngcp-data     253:3    0    5G  0 lvm   /ngcpdata

The /dev/md0 resides in the 3rd partition, where we usually have
plain LVM. /dev/md0 is RAID level 1, on top of the two disks.

Notes:

* we hardcode /dev/md0 as raid-device for mdadm/SW-RAID
* existing data on EFI partitions will be re-used (as grml-debootstrap
  won't overwrite the partition it if contains a filesystem already)
* neither the BIOS (legacy, being first partition) nor the UEFI
  partition (being the second partition) can be part of the SW-RAID
  setup (see https://wiki.debian.org/UEFI -> "RAID for the EFI
  System Partition"). Therefore we need to make sure, that the BIOS
  + UEFI partitions are kept in sync, or at least provide data as
  needed. We will have to handle this during e.g.  upgrades +
  re-installations.
* we need to invoke grub-install for both disks, otherwise
  GRUB is available only from the first disk

+++ templates/scripts/includes/deployment.sh
@@ -67,6 +67,7 @@ ADJUST_FOR_LOW_PERFORMANCE=false
 ENABLE_VM_SERVICES=false
 FILESYSTEM="ext4"
 ROOTFS_SIZE="5G"
+SWRAID_DEVICE=/dev/md0
 GPG_KEY_SERVER="pool.sks-keyservers.net"
 DEBIAN_REPO_HOST="debian.sipwise.com"
 DEBIAN_REPO_TRANSPORT="https"
@@ -372,6 +373,14 @@ if checkBootParam 'targetdisk=' ; then
   TARGET_DISK=$(getBootParam targetdisk)
 fi
 
+if checkBootParam "swraiddisk1=" ; then
+  SWRAID_DISK1=$(getBootParam swraiddisk1)
+fi
+
+if checkBootParam "swraiddisk2=" ; then
+  SWRAID_DISK2=$(getBootParam swraiddisk2)
+fi
+
 # if TARGET_DISK environment variable is set accept it
 if [ -n "$TARGET_DISK" ] ; then
   export DISK="${TARGET_DISK}"
@@ -391,7 +400,17 @@ else # otherwise try to find sane default
   fi
 fi
 
-[ -z "${DISK}" ] && die "Error: No non-removable disk suitable for installation found"
+SWRAID=false
+if [[ -n "${SWRAID_DISK1}" ]] && [[ -z "${SWRAID_DISK2}" ]] ; then
+  die "Error: swraiddisk1 is set, but swraiddisk2 is unset."
+elif [[ -z "${SWRAID_DISK1}" ]] && [[ -n "${SWRAID_DISK2}" ]] ; then
+  die "Error: swraiddisk2 is set, but swraiddisk1 is unset."
+elif [[ -n "${SWRAID_DISK1}" ]] && [[ -n "${SWRAID_DISK2}" ]] ; then
+  echo "Identified valid boot options for Software RAID setup."
+  SWRAID=true
+else
+  [[ -z "${DISK}" ]] && die "Error: No non-removable disk suitable for installation found"
+fi
 
 if checkBootParam 'ngcpstatus=' ; then
   STATUS_WAIT=$(getBootParam ngcpstatus)
@@ -828,6 +847,11 @@ INTERNAL_DEV="${INTERNAL_DEV:-${DEFAULT_INTERNAL_DEV}}"
 if [[ -n "${EXT_GW}" ]]; then
   GW="${EXT_GW}"
 fi
+if [[ "${SWRAID}" = "true" ]] ; then
+  DISK_INFO="Software-RAID with $SWRAID_DISK1 $SWRAID_DISK2"
+else
+  DISK_INFO="/dev/$DISK"
+fi
 
 set_deploy_status "settings"
 
@@ -840,7 +864,7 @@ echo "Deployment Settings:
   $NGCP_INSTALLER_EDITION_STR"
 
 echo "
-  Target disk:       /dev/$DISK
+  Target disk:       $DISK_INFO
   Target Hostname:   $TARGET_HOSTNAME
   Installer version: $SP_VERSION_STR
   Install NW iface:  $INSTALL_DEV
@@ -905,6 +929,9 @@ if "$LOGO" ; then
   else
     EFI_INFO="| no EFI support"
   fi
+  if [[ "${SWRAID}" = "true" ]] ; then
+    SWRAID_INFO="| SW-RAID support [${SWRAID_DISK1} + ${SWRAID_DISK2}]"
+  fi
   # color
   echo -ne "\ec\e[1;32m"
   clear
@@ -916,7 +943,7 @@ if "$LOGO" ; then
   echo "$CPU_INFO CPU(s) | ${RAM_INFO}kB RAM | $CHASSIS"
   echo ""
   echo "Install ngcp: $NGCP_INSTALLER | $INSTALLER_TYPE"
-  echo "Installing $SP_VERSION_STR platform | Debian: $DEBIAN_RELEASE $EFI_INFO $PPA_INFO"
+  echo "Installing $SP_VERSION_STR platform | Debian: $DEBIAN_RELEASE $EFI_INFO $SWRAID_INFO $PPA_INFO"
   echo "Install IP: $INSTALL_IP | Started deployment at $DATE_INFO"
   # number of lines
   echo -ne "\e[10;0r"
@@ -955,8 +982,11 @@ else
 fi
 
 clear_partition_table() {
-  local blockdevice
-  blockdevice="/dev/${DISK}"
+  local blockdevice="$1"
+
+  if [[ ! -b "${blockdevice}" ]] ; then
+    die "Error: ${blockdevice} doesn't look like a valid block device." >&2
+  fi
 
   echo "Wiping disk signatures from ${blockdevice}"
   wipefs -a "${blockdevice}"
@@ -976,15 +1006,12 @@ clear_partition_table() {
     pvremove "$disk" --force --force --yes || true
   done
 
-  dd if=/dev/zero of="/dev/${DISK}" bs=1M count=1
-  blockdev --rereadpt "/dev/${DISK}"
+  dd if=/dev/zero of="${blockdevice}" bs=1M count=1
+  blockdev --rereadpt "${blockdevice}"
 }
 
-set_up_partition_table() {
-  clear_partition_table
-
-  local blockdevice
-  blockdevice="/dev/${DISK}"
+parted_execution() {
+  local blockdevice=$1
 
   echo "Creating partition table"
   parted -a optimal -s "${blockdevice}" mklabel gpt
@@ -999,7 +1026,15 @@ set_up_partition_table() {
   parted -a optimal -s "${blockdevice}" "name 2 'EFI System'"
   parted -a optimal -s "${blockdevice}" set 2 boot on
   EFI_PARTITION="${blockdevice}"2
+}
 
+set_up_partition_table_noswraid() {
+  local blockdevice
+  blockdevice="/dev/${DISK}"
+
+  clear_partition_table "$blockdevice"
+
+  parted_execution "$blockdevice"
   parted -a optimal -s "${blockdevice}" mkpart primary 512M 100%
   parted -a optimal -s "${blockdevice}" "name 3 'Linux LVM'"
   parted -a optimal -s "${blockdevice}" set 3 lvm on
@@ -1008,6 +1043,50 @@ set_up_partition_table() {
   pvcreate -ff -y "${blockdevice}"3
   vgcreate "${VG_NAME}" "${blockdevice}"3
   vgchange -a y "${VG_NAME}"
+}
+
+set_up_partition_table_swraid() {
+  # make sure we don't overlook unassembled SW-RAIDs
+  mdadm --assemble --scan || true # fails if there's nothing to assemble
+
+  if [[ -b "${SWRAID_DEVICE}" ]] ; then
+    echo "NOTE: if you are sure you don't need it SW-RAID device any longer, execute:"
+    echo "      mdadm --remove ${SWRAID_DEVICE} ; mdadm --stop ${SWRAID_DEVICE}; mdadm --zero-superblock /dev/sd..."
+    die "Error: SW-RAID device ${SWRAID_DEVICE} exists already."
+  fi
+
+  for disk in "${SWRAID_DISK1}" "${SWRAID_DISK2}" ; do
+    if grep -q "$disk" /proc/mdstat ; then
+      die "Error: disk $disk seems to be part of an existing SW-RAID setup."
+    fi
+    clear_partition_table "/dev/${disk}"
+  done
+
+  parted_execution "/dev/${SWRAID_DISK1}"
+
+  parted -a optimal -s "/dev/${SWRAID_DISK1}" mkpart primary 512M 100%
+  parted -a optimal -s "/dev/${SWRAID_DISK1}" "name 3 'Linux RAID'"
+  parted -a optimal -s "/dev/${SWRAID_DISK1}" set 3 raid on
+
+  # clone partitioning from SWRAID_DISK1 to SWRAID_DISK2
+  sgdisk "/dev/${SWRAID_DISK1}" -R "/dev/${SWRAID_DISK2}"
+  # randomize the disk's GUID and all partitions' unique GUIDs after cloning
+  sgdisk -G "/dev/${SWRAID_DISK2}"
+
+  echo y | mdadm --create --verbose "${SWRAID_DEVICE}" --level=1 --raid-devices=2 "/dev/${SWRAID_DISK1}"3 "/dev/${SWRAID_DISK2}"3
+
+  echo "Creating PV + VG on ${SWRAID_DEVICE}"
+  pvcreate -ff -y "${SWRAID_DEVICE}"
+  vgcreate "${VG_NAME}" "${SWRAID_DEVICE}"
+  vgchange -a y "${VG_NAME}"
+}
+
+set_up_partition_table() {
+  if [[ "${SWRAID}" = "true" ]] ; then
+    set_up_partition_table_swraid
+  else
+    set_up_partition_table_noswraid
+  fi
 }
 
 create_ngcp_partitions() {
@@ -1063,11 +1142,17 @@ create_debian_partitions() {
 
 display_partition_table() {
   local blockdevice
-  blockdevice="/dev/${DISK}"
-
-  echo "Displaying partition table for reference:"
-  parted -s "${blockdevice}" unit GiB print
-  lsblk "${blockdevice}"
+  if [[ "$SWRAID" = "true" ]] ; then
+    for disk in "${SWRAID_DISK1}" "${SWRAID_DISK2}" ; do
+      echo "Displaying partition table (of SW-RAID disk /dev/$disk) for reference:"
+      parted -s "/dev/${disk}" unit GiB print
+      lsblk "/dev/${disk}"
+    done
+  else
+    echo "Displaying partition table (of /dev/$disk) for reference:"
+    parted -s "/dev/${DISK}" unit GiB print
+    lsblk "/dev/${DISK}"
+  fi
 }
 
 lvm_setup() {
@@ -1119,6 +1204,8 @@ else
 # to be able to retrieve files, starting with Debian/buster no longer present by default
 wget
 EOF
+# required for Software-RAID support
+mdadm
 fi
 
 # NOTE: we use the debian.sipwise.com CNAME by intention here
@@ -1922,8 +2009,34 @@ if [ -r "${INSTALL_LOG}" ] && [ -d "${TARGET}"/var/log/ ] ; then
   cp "${INSTALL_LOG}" "${TARGET}"/var/log/
 fi
 
+if [[ "$SWRAID" = "true" ]] ; then
+  if efi_support ; then
+    chroot "${TARGET}" mount /boot/efi
+
+    echo "Cloning EFI partition from /dev/${SWRAID_DISK1}2 to /dev/${SWRAID_DISK2}2"
+    dd if="/dev/${SWRAID_DISK1}2" of="/dev/${SWRAID_DISK1}2" bs=10M
+    if efi_support ; then
+      if efibootmgr | grep -q 'NGCP Fallback' ; then
+        echo "Deleting existing NGCP Fallback entry from EFI boot manager"
+        efi_entry=$(efibootmgr | awk '/ NGCP Fallback$/ {print $1; exit}' | sed 's/^Boot//; s/\*$//')
+        efibootmgr -b "$efi_entry" -B
+      fi
+      echo "Adding NGCP Fallback entry to EFI boot manager"
+      efibootmgr --create --disk "/dev/${SWRAID_DISK2}" -p 2 -w --label 'NGCP Fallback' --load '\EFI\debian\grubx64.efi'
+    fi
+  fi
+
+  for disk in "${SWRAID_DISK1}" "${SWRAID_DISK2}" ; do
+    grml-chroot "$TARGET" grub-install "/dev/$disk"
+  done
+
+  grml-chroot "$TARGET" update-grub
+fi
+
 # don't leave any mountpoints
 sync
+
+umount ${TARGET}/boot/efi || true
 umount ${TARGET}/proc    || true
 umount ${TARGET}/sys     || true
 umount ${TARGET}/dev/pts || true
